adminapp/requirements.txt
fastapi
uvicorn
jinja2
python-multipart
python-dotenv
asyncpg
aiofiles
requests
pandas
httpx
passlib
apscheduler
